refactor: 가독성을 위한 MoviesAdapter 내부 클래스 분리
