+++ app/src/main/java/woowacourse/movie/feature/movies/view/adapter/AdvertisementViewHolder.kt
@@ -1,0 +1,8 @@
+package woowacourse.movie.feature.movies.view.adapter
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+internal class AdvertisementViewHolder(
+    view: View,
+) : RecyclerView.ViewHolder(view)
